Pass --repo parameter to remote-task.
Currently the run-automated-task.sh script passes the --branch parameter to remote-task,
but there is no way to pass the --repo parameter.

This change lets you set the --repo parameter via TASKS_REPO environment variable,
allowing you to use your own fork of edx-analytics-pipeline.

If TASKS_REPO variable is not defined, it defaults to
https://github.com/edx/edx-analytics-configuration.git