Many tweaks.
Exclude portions from metrics which are not available.  Based on the way
we format the templates with context, if we can't get or don't have
values like the cluster name, or job flow ID, because the caller isn't
running on EMR, we're able to detect and remove those portions from the
metrics.

Also, we now provide a job "index", which allows differentiating
separate runs of the same task within a given cluster run.