[AN-8425] Collect Hadoop metrics after workflow completes
We've added a new Ansible playbook (playfile, really) that will execute
the collection script for us, providing all the necessary bits to SSH to
the proper node.

We've also added a new Make target, which invokes said Ansible playbook,
and we now run it after a workflow completes (after remote-task
returns), before the cluster is terminated.