Add some debugging to metrics collection, which is still tumultuous.
As we're still seeing some weird SSH issues when connecting to the
master node, we're now using -vvvv so we can get the full SSH output as
we try to debug these problems.

We're also invoking the playbook by appending "|| true", so that these
SSH failures don't cause Jenkins to think the job failed.  This is
important because some tasks downstream will not run if their upstream
requirement "failed" because of an SSH error when collecting the
metrics.

We're also spitting out the stdout from the collection script to console
so that we have a better idea of what the script is doing, just in case
that's part of the "SSH error."