Read the workflow logs before collecting metrics.
Some of the workflow jobs generate a lot of log output, and when we're
trying to see if metrics were collected, etc, that stuff is buried
somewhere in the middle of the "console output" for the job.

We're just moving the metrics collection to the very end so that it's
easier to find no matter how large the workflow's log output is.