Allow instance group/instance fleet configuration to peacefully coexist.
Due to how we use/used Ansible, the first iteration of instance fleet
support was being hampered by the default behaviours of Ansible, and by
some of the esoteric syntax we were using.

We slightly changed our logic for detecting fleet vs group, which allows
us to use a more crude-but-effective default value for this work.